refactor out splinter_ref ops also
